feat(localdev): ask for connection in new module wizard (#199)
