+++ src/local-development/create-local-rpc.ts
@@ -1,9 +1,10 @@
 import * as vscode from 'vscode';
 import { AppComponentMetadata } from './types/makecomapp.types';
-import { getMakecomappJson, getMakecomappRootDir } from './makecomappjson';
+import { getMakecomappRootDir } from './makecomappjson';
 import { askFreeText } from './helpers/ask-free-text';
 import { createLocalEmptyComponent } from './create-local-empty-component';
 import { askNewComponentLocalID } from './helpers/ask-component-id';
+import { askForLinkConnection } from './helpers/ask-connection';
 import { catchError } from '../error-handling';
 
 export function registerCommands(): void {
@@ -60,41 +61,4 @@ async function onCreateLocalRpcClick(anyProjectPath: vscode.Uri) {
 	vscode.window.showInformationMessage(
 		`The Remote Procedure "${newRpc.componentLocalId}" sucessfully created locally.`,
 	);
-}
-
-/**
- * Shows the user dialog with a list of available app connections.
- * User can select one, take the option "- no connection -".
- * @throws {Error} if dialog cancelled by user.
- */
-export async function askForLinkConnection(
-	anyProjectPath: vscode.Uri,
-	connectionTypeLabel: string,
-): Promise<string | null | undefined> {
-	const makecomappJson = await getMakecomappJson(anyProjectPath);
-
-	// Prepare dialog options
-	const pickOptions: (vscode.QuickPickItem & { localID: string | null })[] = [
-		// No connection
-		{
-			label: '- no connection -',
-			localID: null,
-		},
-		// All existing connection as options
-		...Object.entries(makecomappJson.components.connection)
-			.filter(([_connectionLocalID, connectionMetadata]) => connectionMetadata !== null)
-			.map(([connectionLocalID, connectionMetadata]) => ({
-				label: `Existing local connection "${connectionMetadata!.label}" [${connectionLocalID}]`,
-				localID: connectionLocalID,
-			})),
-	];
-
-	const componentNamePick = await vscode.window.showQuickPick<(typeof pickOptions)[0]>(pickOptions, {
-		ignoreFocusOut: true,
-		title: `Select the ${connectionTypeLabel}, which to be linked into new local Remote Procedure:`,
-	});
-	if (componentNamePick === undefined) {
-		return undefined;
-	}
-	return componentNamePick.localID;
 }
